backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.2
curl_cffi==0.11.1
distro==1.9.0
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
frozendict==2.4.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
multitasking==0.0.11
numpy==2.2.6
openai==1.82.0
pandas==2.2.3
peewee==3.18.1
pip==23.2.1
platformdirs==4.3.8
protobuf==6.31.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
setuptools==65.5.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.4.0
websockets==15.0.1
yfinance==0.2.61
